+++ src/screens/landing/LandingScreen.tsx
@@ -2,7 +2,7 @@ import { type NativeStackScreenProps } from '@react-navigation/native-stack';
 import { useTranslation } from 'react-i18next';
 import { View, Text, StyleSheet } from 'react-native';
 
-import { Screen, type StackParamList } from '../routing/Router';
+import { Screen, type StackParamList } from '../routing/types';
 
 import { CTA } from '@/components/CTA';
 import { FooterPrompt } from '@/components/sign-in/FooterPrompt';
